WIP: make sticky nav full window, align with content
Goal: the rail of pill components should be able to go all the way to the edge of the screen, but when the screen first loads, it should be aligned with all ofthe other content (even on large screens).

What's in this commit gets us to full window-width, and immediately after first render the rail aligns correctly. But there's a moment on page load where thepills are all the way to the left before they shift.